pyproject.toml
[tool.poetry]
name = "api - cobrança forte"
version = "0.0.4"
description = "Essa API visa trazer uma comunicação de usuário/aplicação de forma eficiente, facilitando a rotina diária referente a controle de pagamentos através do banco CORA.  Desenvolvido para:  ICR ASSESSORIA/FORTE ASSESSORIA. Desenvolvedor BackEnd: Guilherme Novais - GitHub - [GuiNovaisPy, Desenvolvedor BackEnd: Gabriel Silva - GitHub - [GabrielSilvaTech], Desenvolvedor FrontEnd: Luara Oliveira - GitHub - [luluoliv]."
authors = ["GuiNovaisPy , GabrielSilvaTech e luluoliv"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.30"
uvicorn = "^0.30.1"
fastapi = "^0.111.0"
pyodbc = "^5.1.0"
python-dateutil = "^2.9.0.post0"
requests = "^2.32.3"
pyjwt = "^2.8.0"
bcrypt = "^4.1.3"
apscheduler = "^3.10.4"
reportlab = "^4.2.0"
pydantic = "^2.7.3"
bs4 = "^0.0.2"
validate-docbr = "^1.10.0"
alembic = "^1.13.1"
commitizen = "^3.27.0"
mysql-connector-python = "^9.0.0"
numpy = "^2.0.1"
holidays = "^0.54"
paramiko = "^3.4.1"
pandas = "^2.2.2"
openpyxl = "^3.1.5"
geopy = "^2.4.1"
xlsxwriter = "^3.2.0"
selenium = "^4.25.0"
babel = "^2.16.0"
webdriver-manager = "^4.0.2"
pydantic-br = "^1.1.0"
pypdf2 = "^3.0.1"
unidecode = "^1.3.8"
pycryptodome = "^3.21.0"
celery = {extras = ["redis"], version = "^5.4.0"}
flower = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
